refactor: 役割別(levelupと建築量閾値)にtraitを分割
